Refactor score system to use float values
Changed score-related variables and events from int to float in Score, GameDifficulty, and ScoreUI scripts to support fractional scores. Updated ScoreUI to smoothly animate score display using Mathf.Lerp for a more dynamic UI experience.

+++ Assets/Scripts/GameDifficulty.cs
@@ -24,7 +24,7 @@ public class GameDifficulty : MonoBehaviour
         score.OnScoreChanged += HandleScoreChange;
     }
 
-    private void HandleScoreChange(int currentScore)
+    private void HandleScoreChange(float currentScore)
     {
         if (currentScore >= nextThreshold)
         {
